Step towards Q-Learning (SARSA)
From the equation on page 844, we still need to figure out how to get the value for Q(s'a').  Right now, a basic 3D table of state action pairs is defined as a 3D array in the Robot class.  The first two parameters are for the position in the world, and the third is for the action.